+++ tests/Jroc.Test262.Tests/built-ins/Array/prototype/lastIndexOf/ExecutionTests.cs
@@ -1,0 +1,365 @@
+using Jroc.Test262.Tests.built_ins;
+
+namespace Jroc.Test262.Tests.built_ins.Array.prototype.lastIndexOf;
+
+public class ExecutionTests : InMemoryExecutionTestsBase
+{
+    public ExecutionTests() : base("built_ins.Array.prototype.lastIndexOf") { }
+
+    [Fact(DisplayName = "15.4.4.15-1-10")]
+    public Task _15_4_4_15_1_10()
+        => ExecutionTestFromFile("15.4.4.15-1-10");
+
+    [Fact(DisplayName = "15.4.4.15-1-11")]
+    public Task _15_4_4_15_1_11()
+        => ExecutionTestFromFile("15.4.4.15-1-11");
+
+    [Fact(DisplayName = "15.4.4.15-1-12")]
+    public Task _15_4_4_15_1_12()
+        => ExecutionTestFromFile("15.4.4.15-1-12");
+
+    [Fact(DisplayName = "15.4.4.15-1-13")]
+    public Task _15_4_4_15_1_13()
+        => ExecutionTestFromFile("15.4.4.15-1-13");
+
+    [Fact(DisplayName = "15.4.4.15-1-14")]
+    public Task _15_4_4_15_1_14()
+        => ExecutionTestFromFile("15.4.4.15-1-14");
+
+    [Fact(DisplayName = "15.4.4.15-1-15")]
+    public Task _15_4_4_15_1_15()
+        => ExecutionTestFromFile("15.4.4.15-1-15");
+
+    [Fact(DisplayName = "15.4.4.15-1-3")]
+    public Task _15_4_4_15_1_3()
+        => ExecutionTestFromFile("15.4.4.15-1-3");
+
+    [Fact(DisplayName = "15.4.4.15-1-4")]
+    public Task _15_4_4_15_1_4()
+        => ExecutionTestFromFile("15.4.4.15-1-4");
+
+    [Fact(DisplayName = "15.4.4.15-1-5")]
+    public Task _15_4_4_15_1_5()
+        => ExecutionTestFromFile("15.4.4.15-1-5");
+
+    [Fact(DisplayName = "15.4.4.15-1-6")]
+    public Task _15_4_4_15_1_6()
+        => ExecutionTestFromFile("15.4.4.15-1-6");
+
+    [Fact(DisplayName = "15.4.4.15-1-7")]
+    public Task _15_4_4_15_1_7()
+        => ExecutionTestFromFile("15.4.4.15-1-7");
+
+    [Fact(DisplayName = "15.4.4.15-1-8")]
+    public Task _15_4_4_15_1_8()
+        => ExecutionTestFromFile("15.4.4.15-1-8");
+
+    [Fact(DisplayName = "15.4.4.15-1-9")]
+    public Task _15_4_4_15_1_9()
+        => ExecutionTestFromFile("15.4.4.15-1-9");
+
+    [Fact(DisplayName = "15.4.4.15-2-1")]
+    public Task _15_4_4_15_2_1()
+        => ExecutionTestFromFile("15.4.4.15-2-1");
+
+    [Fact(DisplayName = "15.4.4.15-2-10")]
+    public Task _15_4_4_15_2_10()
+        => ExecutionTestFromFile("15.4.4.15-2-10");
+
+    [Fact(DisplayName = "15.4.4.15-2-11")]
+    public Task _15_4_4_15_2_11()
+        => ExecutionTestFromFile("15.4.4.15-2-11");
+
+    [Fact(DisplayName = "15.4.4.15-2-12")]
+    public Task _15_4_4_15_2_12()
+        => ExecutionTestFromFile("15.4.4.15-2-12");
+
+    [Fact(DisplayName = "15.4.4.15-2-13")]
+    public Task _15_4_4_15_2_13()
+        => ExecutionTestFromFile("15.4.4.15-2-13");
+
+    [Fact(DisplayName = "15.4.4.15-2-14")]
+    public Task _15_4_4_15_2_14()
+        => ExecutionTestFromFile("15.4.4.15-2-14");
+
+    [Fact(DisplayName = "15.4.4.15-2-17")]
+    public Task _15_4_4_15_2_17()
+        => ExecutionTestFromFile("15.4.4.15-2-17");
+
+    [Fact(DisplayName = "15.4.4.15-2-18")]
+    public Task _15_4_4_15_2_18()
+        => ExecutionTestFromFile("15.4.4.15-2-18");
+
+    [Fact(DisplayName = "15.4.4.15-2-19")]
+    public Task _15_4_4_15_2_19()
+        => ExecutionTestFromFile("15.4.4.15-2-19");
+
+    [Fact(DisplayName = "15.4.4.15-2-3")]
+    public Task _15_4_4_15_2_3()
+        => ExecutionTestFromFile("15.4.4.15-2-3");
+
+    [Fact(DisplayName = "15.4.4.15-2-5")]
+    public Task _15_4_4_15_2_5()
+        => ExecutionTestFromFile("15.4.4.15-2-5");
+
+    [Fact(DisplayName = "15.4.4.15-2-6")]
+    public Task _15_4_4_15_2_6()
+        => ExecutionTestFromFile("15.4.4.15-2-6");
+
+    [Fact(DisplayName = "15.4.4.15-2-7")]
+    public Task _15_4_4_15_2_7()
+        => ExecutionTestFromFile("15.4.4.15-2-7");
+
+    [Fact(DisplayName = "15.4.4.15-2-9")]
+    public Task _15_4_4_15_2_9()
+        => ExecutionTestFromFile("15.4.4.15-2-9");
+
+    [Fact(DisplayName = "15.4.4.15-3-1")]
+    public Task _15_4_4_15_3_1()
+        => ExecutionTestFromFile("15.4.4.15-3-1");
+
+    [Fact(DisplayName = "15.4.4.15-3-10")]
+    public Task _15_4_4_15_3_10()
+        => ExecutionTestFromFile("15.4.4.15-3-10");
+
+    [Fact(DisplayName = "15.4.4.15-3-11")]
+    public Task _15_4_4_15_3_11()
+        => ExecutionTestFromFile("15.4.4.15-3-11");
+
+    [Fact(DisplayName = "15.4.4.15-3-12")]
+    public Task _15_4_4_15_3_12()
+        => ExecutionTestFromFile("15.4.4.15-3-12");
+
+    [Fact(DisplayName = "15.4.4.15-3-13")]
+    public Task _15_4_4_15_3_13()
+        => ExecutionTestFromFile("15.4.4.15-3-13");
+
+    [Fact(DisplayName = "15.4.4.15-3-14")]
+    public Task _15_4_4_15_3_14()
+        => ExecutionTestFromFile("15.4.4.15-3-14");
+
+    [Fact(DisplayName = "15.4.4.15-3-15")]
+    public Task _15_4_4_15_3_15()
+        => ExecutionTestFromFile("15.4.4.15-3-15");
+
+    [Fact(DisplayName = "15.4.4.15-3-16")]
+    public Task _15_4_4_15_3_16()
+        => ExecutionTestFromFile("15.4.4.15-3-16");
+
+    [Fact(DisplayName = "15.4.4.15-3-17")]
+    public Task _15_4_4_15_3_17()
+        => ExecutionTestFromFile("15.4.4.15-3-17");
+
+    [Fact(DisplayName = "15.4.4.15-3-18")]
+    public Task _15_4_4_15_3_18()
+        => ExecutionTestFromFile("15.4.4.15-3-18");
+
+    [Fact(DisplayName = "15.4.4.15-3-19")]
+    public Task _15_4_4_15_3_19()
+        => ExecutionTestFromFile("15.4.4.15-3-19");
+
+    [Fact(DisplayName = "15.4.4.15-3-2")]
+    public Task _15_4_4_15_3_2()
+        => ExecutionTestFromFile("15.4.4.15-3-2");
+
+    [Fact(DisplayName = "15.4.4.15-3-20")]
+    public Task _15_4_4_15_3_20()
+        => ExecutionTestFromFile("15.4.4.15-3-20");
+
+    [Fact(DisplayName = "15.4.4.15-3-21")]
+    public Task _15_4_4_15_3_21()
+        => ExecutionTestFromFile("15.4.4.15-3-21");
+
+    [Fact(DisplayName = "15.4.4.15-3-23")]
+    public Task _15_4_4_15_3_23()
+        => ExecutionTestFromFile("15.4.4.15-3-23");
+
+    [Fact(DisplayName = "15.4.4.15-3-24")]
+    public Task _15_4_4_15_3_24()
+        => ExecutionTestFromFile("15.4.4.15-3-24");
+
+    [Fact(DisplayName = "15.4.4.15-3-25")]
+    public Task _15_4_4_15_3_25()
+        => ExecutionTestFromFile("15.4.4.15-3-25");
+
+    [Fact(DisplayName = "15.4.4.15-3-28")]
+    public Task _15_4_4_15_3_28()
+        => ExecutionTestFromFile("15.4.4.15-3-28");
+
+    [Fact(DisplayName = "15.4.4.15-3-3")]
+    public Task _15_4_4_15_3_3()
+        => ExecutionTestFromFile("15.4.4.15-3-3");
+
+    [Fact(DisplayName = "15.4.4.15-3-4")]
+    public Task _15_4_4_15_3_4()
+        => ExecutionTestFromFile("15.4.4.15-3-4");
+
+    [Fact(DisplayName = "15.4.4.15-3-5")]
+    public Task _15_4_4_15_3_5()
+        => ExecutionTestFromFile("15.4.4.15-3-5");
+
+    [Fact(DisplayName = "15.4.4.15-3-6")]
+    public Task _15_4_4_15_3_6()
+        => ExecutionTestFromFile("15.4.4.15-3-6");
+
+    [Fact(DisplayName = "15.4.4.15-3-7")]
+    public Task _15_4_4_15_3_7()
+        => ExecutionTestFromFile("15.4.4.15-3-7");
+
+    [Fact(DisplayName = "15.4.4.15-3-9")]
+    public Task _15_4_4_15_3_9()
+        => ExecutionTestFromFile("15.4.4.15-3-9");
+
+    [Fact(DisplayName = "15.4.4.15-4-10")]
+    public Task _15_4_4_15_4_10()
+        => ExecutionTestFromFile("15.4.4.15-4-10");
+
+    [Fact(DisplayName = "15.4.4.15-4-11")]
+    public Task _15_4_4_15_4_11()
+        => ExecutionTestFromFile("15.4.4.15-4-11");
+
+    [Fact(DisplayName = "15.4.4.15-4-2")]
+    public Task _15_4_4_15_4_2()
+        => ExecutionTestFromFile("15.4.4.15-4-2");
+
+    [Fact(DisplayName = "15.4.4.15-4-3")]
+    public Task _15_4_4_15_4_3()
+        => ExecutionTestFromFile("15.4.4.15-4-3");
+
+    [Fact(DisplayName = "15.4.4.15-4-4")]
+    public Task _15_4_4_15_4_4()
+        => ExecutionTestFromFile("15.4.4.15-4-4");
+
+    [Fact(DisplayName = "15.4.4.15-4-5")]
+    public Task _15_4_4_15_4_5()
+        => ExecutionTestFromFile("15.4.4.15-4-5");
+
+    [Fact(DisplayName = "15.4.4.15-4-6")]
+    public Task _15_4_4_15_4_6()
+        => ExecutionTestFromFile("15.4.4.15-4-6");
+
+    [Fact(DisplayName = "15.4.4.15-4-7")]
+    public Task _15_4_4_15_4_7()
+        => ExecutionTestFromFile("15.4.4.15-4-7");
+
+    [Fact(DisplayName = "15.4.4.15-4-8")]
+    public Task _15_4_4_15_4_8()
+        => ExecutionTestFromFile("15.4.4.15-4-8");
+
+    [Fact(DisplayName = "15.4.4.15-4-9")]
+    public Task _15_4_4_15_4_9()
+        => ExecutionTestFromFile("15.4.4.15-4-9");
+
+    [Fact(DisplayName = "15.4.4.15-5-26")]
+    public Task _15_4_4_15_5_26()
+        => ExecutionTestFromFile("15.4.4.15-5-26");
+
+    [Fact(DisplayName = "15.4.4.15-5-27")]
+    public Task _15_4_4_15_5_27()
+        => ExecutionTestFromFile("15.4.4.15-5-27");
+
+    [Fact(DisplayName = "15.4.4.15-8-a-1")]
+    public Task _15_4_4_15_8_a_1()
+        => ExecutionTestFromFile("15.4.4.15-8-a-1");
+
+    [Fact(DisplayName = "15.4.4.15-8-a-11")]
+    public Task _15_4_4_15_8_a_11()
+        => ExecutionTestFromFile("15.4.4.15-8-a-11");
+
+    [Fact(DisplayName = "15.4.4.15-8-a-13")]
+    public Task _15_4_4_15_8_a_13()
+        => ExecutionTestFromFile("15.4.4.15-8-a-13");
+
+    [Fact(DisplayName = "15.4.4.15-8-a-15")]
+    public Task _15_4_4_15_8_a_15()
+        => ExecutionTestFromFile("15.4.4.15-8-a-15");
+
+    [Fact(DisplayName = "15.4.4.15-8-a-2")]
+    public Task _15_4_4_15_8_a_2()
+        => ExecutionTestFromFile("15.4.4.15-8-a-2");
+
+    [Fact(DisplayName = "15.4.4.15-8-a-4")]
+    public Task _15_4_4_15_8_a_4()
+        => ExecutionTestFromFile("15.4.4.15-8-a-4");
+
+    [Fact(DisplayName = "15.4.4.15-8-a-5")]
+    public Task _15_4_4_15_8_a_5()
+        => ExecutionTestFromFile("15.4.4.15-8-a-5");
+
+    [Fact(DisplayName = "15.4.4.15-8-a-7")]
+    public Task _15_4_4_15_8_a_7()
+        => ExecutionTestFromFile("15.4.4.15-8-a-7");
+
+    [Fact(DisplayName = "15.4.4.15-8-a-9")]
+    public Task _15_4_4_15_8_a_9()
+        => ExecutionTestFromFile("15.4.4.15-8-a-9");
+
+    [Fact(DisplayName = "15.4.4.15-8-b-i-1")]
+    public Task _15_4_4_15_8_b_i_1()
+        => ExecutionTestFromFile("15.4.4.15-8-b-i-1");
+
+    [Fact(DisplayName = "15.4.4.15-8-b-i-10")]
+    public Task _15_4_4_15_8_b_i_10()
+        => ExecutionTestFromFile("15.4.4.15-8-b-i-10");
+
+    [Fact(DisplayName = "15.4.4.15-8-b-i-12")]
+    public Task _15_4_4_15_8_b_i_12()
+        => ExecutionTestFromFile("15.4.4.15-8-b-i-12");
+
+    [Fact(DisplayName = "15.4.4.15-8-b-i-14")]
+    public Task _15_4_4_15_8_b_i_14()
+        => ExecutionTestFromFile("15.4.4.15-8-b-i-14");
+
+    [Fact(DisplayName = "15.4.4.15-8-b-i-16")]
+    public Task _15_4_4_15_8_b_i_16()
+        => ExecutionTestFromFile("15.4.4.15-8-b-i-16");
+
+    [Fact(DisplayName = "15.4.4.15-8-b-i-18")]
+    public Task _15_4_4_15_8_b_i_18()
+        => ExecutionTestFromFile("15.4.4.15-8-b-i-18");
+
+    [Fact(DisplayName = "15.4.4.15-8-b-i-19")]
+    public Task _15_4_4_15_8_b_i_19()
+        => ExecutionTestFromFile("15.4.4.15-8-b-i-19");
+
+    [Fact(DisplayName = "15.4.4.15-8-b-i-22")]
+    public Task _15_4_4_15_8_b_i_22()
+        => ExecutionTestFromFile("15.4.4.15-8-b-i-22");
+
+    [Fact(DisplayName = "15.4.4.15-8-b-i-25")]
+    public Task _15_4_4_15_8_b_i_25()
+        => ExecutionTestFromFile("15.4.4.15-8-b-i-25");
+
+    [Fact(DisplayName = "15.4.4.15-8-b-i-26")]
+    public Task _15_4_4_15_8_b_i_26()
+        => ExecutionTestFromFile("15.4.4.15-8-b-i-26");
+
+    [Fact(DisplayName = "15.4.4.15-8-b-i-27")]
+    public Task _15_4_4_15_8_b_i_27()
+        => ExecutionTestFromFile("15.4.4.15-8-b-i-27");
+
+    [Fact(DisplayName = "15.4.4.15-8-b-i-29")]
+    public Task _15_4_4_15_8_b_i_29()
+        => ExecutionTestFromFile("15.4.4.15-8-b-i-29");
+
+    [Fact(DisplayName = "15.4.4.15-8-b-i-4")]
+    public Task _15_4_4_15_8_b_i_4()
+        => ExecutionTestFromFile("15.4.4.15-8-b-i-4");
+
+    [Fact(DisplayName = "15.4.4.15-8-b-i-6")]
+    public Task _15_4_4_15_8_b_i_6()
+        => ExecutionTestFromFile("15.4.4.15-8-b-i-6");
+
+    [Fact(DisplayName = "15.4.4.15-8-b-i-8")]
+    public Task _15_4_4_15_8_b_i_8()
+        => ExecutionTestFromFile("15.4.4.15-8-b-i-8");
+
+    [Fact(DisplayName = "15.4.4.15-9-2")]
+    public Task _15_4_4_15_9_2()
+        => ExecutionTestFromFile("15.4.4.15-9-2");
+
+    [Fact(DisplayName = "call-with-boolean")]
+    public Task call_with_boolean()
+        => ExecutionTestFromFile("call-with-boolean");
+
+}