+++ tests/Jroc.Test262.Tests/built-ins/Array/prototype/pop/ExecutionTests.cs
@@ -15,4 +15,45 @@ public class ExecutionTests : InMemoryExecutionTestsBase
     [Fact(DisplayName = "S15.4.4.6_A1.2_T1")]
     public Task S15_4_4_6_A1_2_T1()
         => ExecutionTestFromFile("S15.4.4.6_A1.2_T1");
+
+    [Fact(DisplayName = "S15.4.4.6_A2_T1")]
+    public Task S15_4_4_6_A2_T1()
+        => ExecutionTestFromFile("S15.4.4.6_A2_T1");
+
+    [Fact(DisplayName = "S15.4.4.6_A2_T2")]
+    public Task S15_4_4_6_A2_T2()
+        => ExecutionTestFromFile("S15.4.4.6_A2_T2");
+
+    [Fact(DisplayName = "S15.4.4.6_A2_T3")]
+    public Task S15_4_4_6_A2_T3()
+        => ExecutionTestFromFile("S15.4.4.6_A2_T3");
+
+    [Fact(DisplayName = "S15.4.4.6_A2_T4")]
+    public Task S15_4_4_6_A2_T4()
+        => ExecutionTestFromFile("S15.4.4.6_A2_T4");
+
+    [Fact(DisplayName = "S15.4.4.6_A3_T1")]
+    public Task S15_4_4_6_A3_T1()
+        => ExecutionTestFromFile("S15.4.4.6_A3_T1");
+
+    [Fact(DisplayName = "S15.4.4.6_A3_T2")]
+    public Task S15_4_4_6_A3_T2()
+        => ExecutionTestFromFile("S15.4.4.6_A3_T2");
+
+    [Fact(DisplayName = "S15.4.4.6_A3_T3")]
+    public Task S15_4_4_6_A3_T3()
+        => ExecutionTestFromFile("S15.4.4.6_A3_T3");
+
+    [Fact(DisplayName = "S15.4.4.6_A4_T1")]
+    public Task S15_4_4_6_A4_T1()
+        => ExecutionTestFromFile("S15.4.4.6_A4_T1");
+
+    [Fact(DisplayName = "S15.4.4.6_A4_T2")]
+    public Task S15_4_4_6_A4_T2()
+        => ExecutionTestFromFile("S15.4.4.6_A4_T2");
+
+    [Fact(DisplayName = "call-with-boolean")]
+    public Task call_with_boolean()
+        => ExecutionTestFromFile("call-with-boolean");
+
 }